% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_ojp_time}
\alias{get_ojp_time}
\title{get_ojp_time get ...}
\usage{
get_ojp_time(
  auth = token1,
  origin = NA,
  destination = NA,
  time = "2020-12-16T08:34:40"
)
}
\arguments{
\item{auth}{authentication token}

\item{origin}{geocoordinates of the origin, WGS}

\item{destination}{geocoordinates of the destination}

\item{time}{departure time in format "YYYY-MM-16T08:34:40"}
}
\value{
tibble
}
\description{
get_ojp_time get ...
}
\examples{
coords <- geo(address = c("Zürich Schwamendingerplatz, Switzerland", 
                          "Bahnhof Stettbach, Switzerland"), method ='osm')

ojp <- get_ojp_time(
  origin=c(coords[1,]$long,coords[1,]$lat), 
  destination=c(coords[2,]$long,coords[2,]$lat)
)
}
