% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tripduration.R
\name{get_tripduration}
\alias{get_tripduration}
\title{Get tripduration}
\usage{
get_tripduration(
  auth = NA,
  origin = NA,
  destination = NA,
  time = NA,
  sys.sleep = NA
)
}
\arguments{
\item{auth}{authentication token}

\item{origin}{geocoordinates of the origin, WGS}

\item{destination}{geocoordinates of the destination}

\item{time}{departure time in format "YYYY-MM-16T08:34:40"}

\item{sys.sleep}{delay after an API-call until the next is triggered}
}
\value{
tibble
}
\description{
Get tripduration
}
\examples{

origin <- list(c(8.572164,47.40459),
               c(8.59557,47.39675))

destination <- c(7.45,46.9)

get_tripduration(auth="your_token", origin=origin,destination,time, sys.sleep=2)

}
